Extract SharedPreferences data store implementation
