--- sdk/core/src/main/java/com/klaviyo/coresdk/utils/DataStoreInterface.kt
@@ -1,10 +1,0 @@
-package com.klaviyo.coresdk.utils
-
-/**
- * TODO asynchronous?
- */
-internal interface DataStoreInterface {
-    fun fetch(key: String): String?
-
-    fun store(key: String, value: String)
-}
